#177 refactor: 와인 강의 완료 리팩토링 - get메소드가 존재하지 않아서 dsl은 인터페이스까지 생성 - 수강완료 삭제 인자를 기존 wineLectureCompleteId에서 wineLectureId와 memberId를 기준으로 변경 - 유저의 완료한 강의 보기 기능 삭제
