+++ src/main/java/com/drinkeg/drinkeg/repository/wineLectureComplete/WineLectureCompleteRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.drinkeg.drinkeg.repository.wineLectureComplete;
+
+public interface WineLectureCompleteRepositoryCustom {
+}
